fixed libtool

+++ var/spack/repos/scs_io/packages/libtool/package.py
@@ -60,8 +60,10 @@ class Libtool(AutotoolsPackage, GNUMirrorPackage):
         # GNU commands glibtool and glibtoolize instead of the BSD
         # variant; this happens frequently, for instance, on Darwin
         import os
-        os.remove(join_path(self.prefix.bin, 'glibtool'))
-        os.remove(join_path(self.prefix.bin, 'glibtoolize'))
+        if os.path.exists(join_path(self.prefix.bin, 'glibtool')):
+            os.remove(join_path(self.prefix.bin, 'glibtool'))
+        if os.path.exists(join_path(self.prefix.bin, 'glibtoolize')):
+            os.remove(join_path(self.prefix.bin, 'glibtoolize'))
         symlink(join_path(self.prefix.bin, 'libtool'),
                 join_path(self.prefix.bin, 'glibtool'))
         symlink(join_path(self.prefix.bin, 'libtoolize'),
